+++ src/Header/AcceptRanges.php
@@ -3,44 +3,67 @@ declare(strict_types = 1);
 
 namespace Innmind\Http\Header;
 
-use Innmind\Http\Header as HeaderInterface;
-use Innmind\Immutable\Sequence;
+use Innmind\Http\{
+    Header as HeaderInterface,
+    Exception\DomainException,
+};
+use Innmind\Immutable\{
+    Sequence,
+    Str,
+    Maybe,
+};
 
 /**
  * @psalm-immutable
  */
 final class AcceptRanges implements HeaderInterface
 {
-    private Header $header;
-
-    public function __construct(AcceptRangesValue $ranges)
-    {
-        $this->header = new Header('Accept-Ranges', $ranges);
+    private function __construct(
+        private string $ranges,
+    ) {
     }
 
     /**
      * @psalm-pure
+     *
+     * @throws DomainException
      */
     public static function of(string $range): self
     {
-        return new self(new AcceptRangesValue($range));
+        return self::maybe($range)->match(
+            static fn($self) => $self,
+            static fn() => throw new DomainException($range),
+        );
+    }
+
+    /**
+     * @psalm-pure
+     *
+     * @return Maybe<self>
+     */
+    public static function maybe(string $range): Maybe
+    {
+        return Maybe::just($range)
+            ->map(Str::of(...))
+            ->filter(static fn($range) => $range->matches('~^\w+$~'))
+            ->map(static fn() => new self($range));
     }
 
     #[\Override]
     public function name(): string
     {
-        return $this->header->name();
+        return 'Accept-Ranges';
     }
 
     #[\Override]
     public function values(): Sequence
     {
-        return $this->header->values();
+        return Sequence::of(new Value\Value($this->ranges));
     }
 
     #[\Override]
     public function toString(): string
     {
-        return $this->header->toString();
+        return (new Header($this->name(), ...$this->values()->toList()))->toString();
     }
 }