+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - `Innmind\Http\Factory\HeadersFactory` is now a final class
 - Classes in `Innmind\Http\Factory\Header\` are now internal
 - `Innmind\Http\Header::values()` now returns an `Innmind\Immutable\Sequence` to allow for ordered values
+- `Innmind\Http\Header\AcceptRanges` constructor is now private, use `::of()` or `::maybe()` named constructors
 
 ### Removed
 
@@ -28,6 +29,7 @@
 - `Innmind\Http\Factory\ServerRequest\ServerRequestFactory`
 - `Innmind\Http\Factory\Header\HeadersFactory`
 - `Innmind\Http\Factory\HeaderFactory`
+- `Innmind\Http\Header\AcceptRangesValue`
 
 ### Fixed